Use iroh 0.34 (#48)
* Use latest fendermint image that has iroh 0.34 as a library.
* Remove iroh service (because it is embedded into fendermint and
objects services).
* Add `docker_network_subnet` configuration parameter
* Assign static IPs to all node containers based on
`docker_network_subnet`. This simplifies IP resolution for iroh and also
potentially for HTTP LBs, e.g. caddy can be deployed as a host service
and forward requests to a specific cometbft IP. Before this PR, caddy
needs to run as a container in the same docker network to be able to
route requests to docker DNS names.

Close #46